scalable-version/streamlit_app/requirements.txt
streamlit
requests
pillow
flask
textwrap3